first param in Vertex constructor casted to double

+++ polygon_backend/src/main/java/com/atis/polygon_area/geometry/Vector.java
@@ -7,9 +7,9 @@ import lombok.Data;
 @AllArgsConstructor
 
 public class Vector {
-    private float X;
-    private float Y;
-    private float Z;
+    private double X;
+    private double Y;
+    private double Z;
 
     public Vector(Vertex v) {
         this.X = v.getX();
